+++ src/lang/sem/rel/mod.rs
@@ -1,22 +1,24 @@
 // devela/src/lang/sem/rel/mod.rs
 //
-//! Abstract relational semantics.
+//! Semantic relations.
 //!
-//! > How meaning can relate.
+//! > How meaning relates.
 //
 
-// mod about; // aboutness
-// mod comp; // composed relations
-// mod equiv; // semantic sameness
-// mod logic; // formal relations
+// mod about; // Aboutness and semantic reference
+// mod comp; // Composition of relations
+// mod equiv; // Semantic equivalence
+mod relation; // Relation
+// mod sim; // Semantic similarity and affinity
 
 crate::structural_mods! { // _mods
     _mods {
-        // pub use super::{
-        //     about::_all::*,
-        //     comp::_all::*,
-        //     equiv::_all::*,
-        //     logic::_all::*,
-        // };
+        pub use super::{
+            // about::_all::*,
+            // comp::_all::*,
+            // equiv::_all::*,
+            relation::Relation,
+            // sim::_all::*,
+        };
     }
 }